refactor: move `__sys_{free,malloc,realloc}` to `syscalls`
